requirements.txt
pyyaml==6.0.1
python-frontmatter==1.1.0